maxspeed: Move into folder, minor changes, add docs, add processing
