+++ app/process/maxspeed/README.md
@@ -1,0 +1,23 @@
+# Dataset `maxspeed`
+
+**OSM:**
+
+- Docs: https://wiki.openstreetmap.org/wiki/Key:maxspeed
+
+## Main values
+
+- TODO
+
+## Style helper tags
+
+_None_
+
+## Raw OSM tags
+
+(!) We might change those tags at any time.
+
+> TODO
+
+## Processing helper
+
+_None_
